+++ scripts/validate_all_mmcif_worker.sh
@@ -125,18 +125,22 @@ do
   lock_file=$WORK_DIR/$entry_id.lock
   parser_log=$cif_label-parser.log
 
-  touch $lock_file
+  if [ ! -e $lock_file ] ; then
 
-  rm -f $cif_dir/$diag_log $cif_dir/$parser_log
+   touch $lock_file
 
-  ( cd $cif_dir ; CifCheck -f $_cif_file -dictSdb $dict_sdb > /dev/null ; [ -e $diag_log ] && [ `LC_ALL=C grep -F -v 'has invalid value "?" in row' $diag_log | sed -e /^$/d | wc -l 2> /dev/null` = 0 ] && rm -f $diag_log )
-  ( cd $cif_dir ; [ ! -e $diag_log ] && [ ! -e $parser_log ] && echo $new_chk_sum > $chk_sum_file ; [ -e $parser_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $parser_log ) ; [ -e $diag_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $diag_log ) )
+   rm -f $cif_dir/$diag_log $cif_dir/$parser_log
 
-  if [ $proc_id_mod -eq 0 ] ; then
-   echo -e -n "\rDone "$((proc_id + 1)) of $TOTAL ...
+   ( cd $cif_dir ; CifCheck -f $_cif_file -dictSdb $dict_sdb > /dev/null ; [ -e $diag_log ] && [ `LC_ALL=C grep -F -v 'has invalid value "?" in row' $diag_log | sed -e /^$/d | wc -l 2> /dev/null` = 0 ] && rm -f $diag_log )
+   ( cd $cif_dir ; [ ! -e $diag_log ] && [ ! -e $parser_log ] && echo $new_chk_sum > $chk_sum_file ; [ -e $parser_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $parser_log ) ; [ -e $diag_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $diag_log ) )
+
+   if [ $proc_id_mod -eq 0 ] ; then
+    echo -e -n "\rDone "$((proc_id + 1)) of $TOTAL ...
+   fi
+
+   rm -f $lock_file
+
   fi
-
-  rm -f $lock_file
 
  fi
 
@@ -202,10 +206,14 @@ do
 
   if [ ! -e $lock_file ] ; then
 
-  rm -f $cif_dir/$diag_log $cif_dir/$parser_log
+   touch $lock_file
 
-  ( cd $cif_dir ; CifCheck -f $_cif_file -dictSdb $dict_sdb > /dev/null ; [ -e $diag_log ] && [ `LC_ALL=C grep -F -v 'has invalid value "?" in row' $diag_log | sed -e /^$/d | wc -l 2> /dev/null` = 0 ] && rm -f $diag_log )
-  ( cd $cif_dir ; [ ! -e $diag_log ] && [ ! -e $parser_log ] && echo $new_chk_sum > $chk_sum_file ; [ -e $parser_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $parser_log ) ; [ -e $diag_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $diag_log ) )
+   rm -f $cif_dir/$diag_log $cif_dir/$parser_log
+
+   ( cd $cif_dir ; CifCheck -f $_cif_file -dictSdb $dict_sdb > /dev/null ; [ -e $diag_log ] && [ `LC_ALL=C grep -F -v 'has invalid value "?" in row' $diag_log | sed -e /^$/d | wc -l 2> /dev/null` = 0 ] && rm -f $diag_log )
+   ( cd $cif_dir ; [ ! -e $diag_log ] && [ ! -e $parser_log ] && echo $new_chk_sum > $chk_sum_file ; [ -e $parser_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $parser_log ) ; [ -e $diag_log ] && ( [ $DELETE = "true" ] && rm -f $cif_file ; cat $diag_log ) )
+
+   rm -f $lock_file
 
   fi
